Remove graph_size.mk when updating
In 2.2, we wrote invalid data to this file, hence, we remove it when
updating.

+++ cmk/update_config/plugins/actions/remove_persisted_graph_options.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
-# Copyright (C) 2019 Checkmk GmbH - License: GNU General Public License v2
+# Copyright (C) 2023 Checkmk GmbH - License: GNU General Public License v2
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
+from itertools import chain
 from logging import Logger
 
 from cmk.utils.paths import profile_dir
@@ -11,16 +12,26 @@ from cmk.update_config.registry import update_action_registry, UpdateAction
 from cmk.update_config.update_state import UpdateActionState
 
 
-class RemovePersistedGraphRanges(UpdateAction):
+class RemovePersistedGraphOptions(UpdateAction):
+    _KEY = "persisted_graph_options_removed"
+
     def __call__(self, logger: Logger, update_action_state: UpdateActionState) -> None:
-        for path in profile_dir.glob("*/graph_range.mk"):
+        if update_action_state.get(self._KEY):
+            return
+
+        for path in chain(
+            profile_dir.glob("*/graph_range.mk"),
+            profile_dir.glob("*/graph_size.mk"),
+        ):
             path.unlink()
+
+        update_action_state[self._KEY] = "True"
 
 
 update_action_registry.register(
-    RemovePersistedGraphRanges(
-        name="remove_persisted_graph_ranges",
-        title="Remove persisted graph ranges",
+    RemovePersistedGraphOptions(
+        name="remove_persisted_graph_options",
+        title="Remove persisted graph options",
         sort_index=150,
     )
 )
